Decouple strict breaks and postponing breaks.
Addresses issue #316.

Because there's already separate options for skipping and postponing breaks, this PR decouples skips and postpones. Previously, if strict break was enabled, there would be no way to postpone the break even if the user had "allow postponing" on. This PR allows the user to choose to postpone but not skip breaks in the settings menu.