Remove use of distutils
As of Python 3.12 setuptools is no longer installed by default and
if setuptools is not installed distutils is not found and SafeEyes
fails to start.

Furthermore distutils is deprecated and should thus not be used
anymore. The packaging module provides a replacement for LooseVersion
and it is also recommended by PEP 632 [1].

Also update required Python version for Debian to 3.12 or newer.

[1] https://peps.python.org/pep-0632/#migration-advice