rework custom stylesheet loading
Previously, the stylesheet was copied to the user's .config dir, and
this was then loaded as a style (with priority "application").
The user could alter this stylesheet directly.
This stylesheet was called "safeeyes_styles.css".

Now, we split the stylesheets into two:
- The system stylesheet, which is directly loaded with priority
  "application".
- The user's stylesheet, if it exists, which is loaded from the .config
  dir with priority "user", overriding both the system stylesheet and
  any themes the user might have.
  This is now called "safeeyes_custom_styles.css".

This means in the future, we can simply add new default styles to the
system stylesheet without breaking the user. Additionally, any styles
loaded with a higher priority than "application" (eg. certain GTK
themes, see https://github.com/slgobinath/SafeEyes/issues/687), can now
be overridden by the user's stylesheet.

We also check if the old path ("safeeyes_styles.css") contains an old
default stylesheet. If it does, we delete it.
If it (likely) contains a customized stylesheet, don't delete it, but
warn the user that they should use "safeeyes_custom_styles.css" instead
now.