trayicon: do not block main thread with animation
The plugin hooks (on_pre_break, on_start_break) are always called on the
main thread (and most plugins rely on that, so this is pretty much
guaranteed at this point).
Before this commit, start_animation was blocking (`time.sleep()`) for a
second in total, blocking the main thread. Only the second call to
start_animation() happened in a separate thread.
This can be avoided by using GLib timeouts instead of blocking. We also
save the extra thread (and any synchronization it'd need), and a bunch
of execute_main_thread calls.