core: add __fire_hook method
this adds a __fire_hook method for safely firing an event hook on the
main thread, and uses it to fire all hooks.
this ensures that all hooks are always fired from the main thread in one
central place. this makes changing that logic easier later
this also fixes the on_count_down hook, which previously wasn't always
called from the main thread.
__fire_hook also ensures that plugins don't do anything weird like
recursively calling methods that would trigger another hook instantly.

the implementation of __fire_hook is somewhat clever. it uses a
threading.Event to block on and synchronize with the event loop, so we
can wait for the plugins' response and return it.
since this implementation will later be replaced, it is okay that it is
somewhat complicated.