core: refactor: add _take_break_now flag
before this, the implementation is complicated and slow.
it stops the scheduler completely, waits for a second, and then restarts
it at the correct place.
this change removes all that and replaces it with setting a
_take_break_now flag and waking the scheduler. this flag is checked
after every __wait_for call and, if applicable, calls the correct code.

this makes taking a break using the -t cli flag or the "Take break now"
option noticeably faster. it also clears up the code, making it very
clear when and where a break can be taken.